requirements.txt
# MarketForge AI — Backend Production Dependencies
# Stripped of Airflow / MLflow / ChromaDB / heavy training deps.
# APScheduler replaces Airflow. Railway provides PostgreSQL + Redis.

# ── Agent Orchestration ──────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langsmith>=0.1.0

# ── LLM & Embeddings ────────────────────────────────────────────────────────
google-generativeai>=0.8.0
sentence-transformers>=3.0.0

# ── NLP ─────────────────────────────────────────────────────────────────────
spacy>=3.7.0
flashtext>=2.7
rank-bm25>=0.2.2

# ── ML (inference only — no training) ───────────────────────────────────────
scikit-learn>=1.4.0
lightgbm>=4.3.0

# ── Web / Scraping ───────────────────────────────────────────────────────────
tavily-python>=0.5.0
httpx>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=5.2.0

# ── Data ────────────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
pandas>=2.2.0
numpy>=1.26.0

# ── Persistence ──────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
redis>=5.0.0

# ── API ──────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Scheduler (replaces Airflow) ─────────────────────────────────────────────
apscheduler>=3.10.0

# ── Security ─────────────────────────────────────────────────────────────────
transformers>=4.41.0
torch>=2.3.0

# ── Observability ─────────────────────────────────────────────────────────────
structlog>=24.1.0
prometheus-client>=0.20.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tenacity>=8.3.0
tiktoken>=0.7.0
python-dateutil>=2.9.0
pytz>=2024.1
